wip changing auth to promises
